Move Rosenblum's feature templates to Method
Rosenblum's feature extraction methods have been moved to the "Feature
extraction" section of Method, which (I think) really helps with the flow
of the thesis. Some technical details such as the construction of the
CCFG have been moved to the technical background as well. Rosenblum's
section in the state-of-the-art analysis will probably have to be
rewritten after this.